+++ rxbinding-appcompat-v7/src/androidTest/java/com/jakewharton/rxbinding2/support/v7/widget/RxSearchViewTest.java
@@ -1,4 +1,4 @@
-package com.jakewharton.rxbinding.support.v7.widget;
+package com.jakewharton.rxbinding2.support.v7.widget;
 
 import android.content.Context;
 import android.support.test.InstrumentationRegistry;
@@ -7,8 +7,8 @@ import android.support.test.rule.UiThreadTestRule;
 import android.support.test.runner.AndroidJUnit4;
 import android.support.v7.widget.SearchView;
 import android.view.ContextThemeWrapper;
-import com.jakewharton.rxbinding.RecordingObserver;
-import com.jakewharton.rxbinding.support.v7.appcompat.R;
+import com.jakewharton.rxbinding2.RecordingObserver;
+import com.jakewharton.rxbinding2.support.v7.appcompat.R;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
