+++ rxbinding-appcompat-v7/src/main/AndroidManifest.xml
@@ -1,1 +1,1 @@
-<manifest package="com.jakewharton.rxbinding.support.v7.appcompat"/>
+<manifest package="com.jakewharton.rxbinding2.support.v7.appcompat"/>
