+++ rxbinding-appcompat-v7/src/main/java/com/jakewharton/rxbinding2/support/v7/widget/PopupMenuDismissObservable.java
@@ -1,4 +1,4 @@
-package com.jakewharton.rxbinding.support.v7.widget;
+package com.jakewharton.rxbinding2.support.v7.widget;
 
 import android.support.v7.widget.PopupMenu;
 import com.jakewharton.rxbinding2.internal.Notification;
