+++ rxbinding-appcompat-v7/src/main/java/com/jakewharton/rxbinding2/support/v7/widget/SearchViewQueryTextChangesObservable.java
@@ -1,4 +1,4 @@
-package com.jakewharton.rxbinding.support.v7.widget;
+package com.jakewharton.rxbinding2.support.v7.widget;
 
 import android.support.v7.widget.SearchView;
 import io.reactivex.Observable;
@@ -7,45 +7,41 @@ import io.reactivex.android.MainThreadDisposable;
 
 import static io.reactivex.android.MainThreadDisposable.verifyMainThread;
 
-final class SearchViewQueryTextChangeEventsObservable extends Observable<SearchViewQueryTextEvent> {
-  private final SearchView view;
+final class SearchViewQueryTextChangesObservable extends Observable<CharSequence> {
+  final SearchView view;
 
-  SearchViewQueryTextChangeEventsObservable(SearchView view) {
+  SearchViewQueryTextChangesObservable(SearchView view) {
     this.view = view;
   }
 
-  @Override protected void subscribeActual(Observer<? super SearchViewQueryTextEvent> observer) {
+  @Override protected void subscribeActual(Observer<? super CharSequence> observer) {
     verifyMainThread();
     Listener listener = new Listener(view, observer);
     observer.onSubscribe(listener);
     view.setOnQueryTextListener(listener);
 
     // Emit initial value.
-    observer.onNext(SearchViewQueryTextEvent.create(view, view.getQuery(), false));
+    observer.onNext(view.getQuery());
   }
 
   final class Listener extends MainThreadDisposable implements SearchView.OnQueryTextListener {
     private final SearchView searchView;
-    private final Observer<? super SearchViewQueryTextEvent> observer;
+    private final Observer<? super CharSequence> observer;
 
-    Listener(SearchView searchView, Observer<? super SearchViewQueryTextEvent> observer) {
+    Listener(SearchView searchView, Observer<? super CharSequence> observer) {
       this.searchView = searchView;
       this.observer = observer;
     }
 
     @Override public boolean onQueryTextChange(String s) {
       if (!isDisposed()) {
-        observer.onNext(SearchViewQueryTextEvent.create(view, s, false));
+        observer.onNext(s);
         return true;
       }
       return false;
     }
 
     @Override public boolean onQueryTextSubmit(String query) {
-      if (!isDisposed()) {
-        observer.onNext(SearchViewQueryTextEvent.create(view, view.getQuery(), true));
-        return true;
-      }
       return false;
     }
